add method for sidebar+refactor
